# src/public/openapi.yml
openapi: 3.0.3
info:
  version: "1"
  title: "Express Boilerplate"
  description: "Demo Express Boilerplate"
paths:
  /auth/sign-up:
    post:
      summary: Sign Up User
      tags:
        - UserApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUp"
      responses:
        200:
          $ref: "#/components/responses/Response"
  /auth/sign-in:
    post:
      summary: Sign In User
      tags:
        - UserApi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignIn"
      responses:
        200:
          $ref: "#/components/responses/Response"

    get:
      summary: Get all users
      tags:
        - UserApi
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
    delete:
      summary: Delete all users
      operationId: deleteAllUsers
      responses:
        204:
          description: Success
...
components:
  responses:
    Response:
      description: Common response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
            properties:
              status:
                type: string
                enum: [ok, error]
              code:
                type: number
                example:
                  - 200
              message:
                type: string
                enum:
                  - Success
                  - Request body has wrong format
...
  schemas:
    UserSignUp:
      description: UserSignUp
      type: object
      required:
        - email
        - password
        - confirmPassword
      properties:
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
      example:
        email: test@email.com
        password: userpassword
        confirmPassword: userpassword
    UserSignIn:
      description: UserSignIn
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: test@email.com
        password: userpassword
...
    User:
      example: {
        tokens: [],
        _id: "5eb9a1968b8d5b36155290ac",
        email: "test@email.com",
        password: "$2a$10$kcq5VxeRO7ZS01FsUAqshOG.6b9czvW27beI6jSfsX7CHdYR4cL2m",
        createdAt: "2020-05-11T19:03:50.581Z",
        updatedAt: "2020-05-11T19:03:50.581Z",
        __v: 0
      }

    # ArrayOfBooks #
...
    Users:
      description: List of Users
      type: array
      items:
        $ref: '#/components/schemas/User'